refactor: General backend and frontend changes and improvements - add one off payments, products and orders domain - refine email send methods, and subscription event methods - add suspense loading, seprate list components and improve typescript type definition

+++ backend/src/main/java/com/stripe/payment_service_provider/subscription/repository/PlanRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PlanRepository extends JpaRepository<StripePlan, Integer> {
+public interface PlanRepository extends JpaRepository<StripePlan, Long> {
     Optional<StripePlan> findSubscriptionPlanByStripeProductId(String id);
 }
